requirements.txt
numpy>=1.19.0
Pillow>=8.0.0
simpleaudio>=1.0.0